requirements.txt
langchain==0.1.16
langchain_openai==0.1.3
python-dotenv==1.0.1
fastapi==0.110.2
uvicorn==0.29.0
langchainhub==0.1.15
unstructured[pdf]
faiss-cpu
ipykernel
langchain_chroma==0.1.1
ipython